Refactor into package for addition of setup; remove pcp attr of Problem
