+++ pyrallelsa/examples/tsp/__init__.py
@@ -1,10 +1,10 @@
-import random
-import math
 import json
+import math
+import random
+
 import jsonschema
 
-from pyrallelsa import Annealer
-from pyrallelsa import ProblemClassPath
+from pyrallelsa import Problem
 
 
 def distance(a, b):
@@ -28,7 +28,7 @@ def get_distance_matrix(cities):
     return distance_matrix
 
 
-class TSPProblem(Annealer):
+class TSPProblem(Problem):
     """Traveling Salesman Problem Annealer
 
     :param State|None state: state of the current annealer process; if this
@@ -100,10 +100,6 @@ class TSPProblem(Annealer):
             for i in range(len(route)):
                 e += distance(self.cities[route[i-1]], self.cities[route[i]])
         return e
-
-    @classmethod
-    def pcp(self):
-        return ProblemClassPath("tsp", "TSPProblem")
 
     @classmethod
     def divide(cls, divisions, problem_data):
